requirements.txt
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.0
#jupyter>=1.0.0
python-dotenv>=0.19.0
basedosdados>=2.0.3
rich>=14.3.3